feat(monorepo): Phase 2 — extract @mulmobridge/chat-service (#360)
server/api/chat-service/ → packages/chat-service/ with zero
cross-boundary imports remaining.

### Package: @mulmobridge/chat-service

DI-pure factory — all host-app concerns injected via ChatServiceDeps.
Provides socket.io + REST bridge to Claude Code agents.

- Source: packages/chat-service/src/ (7 files)
- Tests: packages/chat-service/test/ (3 files, 22 tests)
- README: install, usage, architecture diagram, exports table
- Dependencies: @mulmobridge/protocol, express (peer), socket.io (peer)

### DI fix

chat-state.ts: writeFileAtomic (server util dependency) → writeFile
from fs/promises. Removes the last cross-boundary import.

### Shared config

- config/tsconfig.packages.json — shared base for all @mulmobridge/*
- config/eslint.packages.mjs — shared ESLint base (tseslint + prettier + sonarjs)
- Per-package eslint.config.mjs extends the base

### Import migration

- server/index.ts: ./api/chat-service → @mulmobridge/chat-service
- server/agent/*.ts: Attachment type → @mulmobridge/protocol
- test/chat-service/ → packages/chat-service/test/ (relative imports)
