+++ packages/chat-service/src/commands.ts
