+++ packages/protocol/eslint.config.mjs
@@ -1,0 +1,7 @@
+import eslintBase from "../../config/eslint.packages.mjs";
+
+export default [
+  { files: ["{src,test}/**/*.ts"] },
+  { ignores: ["dist/**/*"] },
+  ...eslintBase,
+];
